move Commands to features
