+++ src/pages/content/Providers.tsx
@@ -1,7 +1,7 @@
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
 import "@src/assets/styles/studres-css-reset.css"
 import "@src/assets/styles/shadcn-ui.css"
-import { CommandProvider } from "@src/components/command/CommandContext"
+import { CommandProvider } from "@src/features/command/CommandContext"
 import { TooltipProvider } from "@src/components/ui/tooltip"
 import { SmoothRouterListener } from "@src/features/router/SmoothRouterListener"
 import { PreferredTheme, ThemeProvider } from "@src/features/theme"
