+++ src/pages/content/pages/Folder.tsx
@@ -1,7 +1,7 @@
 import parse from "html-react-parser"
-import { useCommand } from "@src/components/command/CommandContext"
-import CommandInput from "@src/components/command/CommandInput"
-import Commands from "@src/components/command/Commands"
+import { useCommand } from "@src/features/command/CommandContext"
+import CommandInput from "@src/features/command/CommandInput"
+import Commands from "@src/features/command/Commands"
 import CompactLayout from "@src/components/layouts/CompactLayout"
 import MainLayout from "@src/components/layouts/MainLayout"
 import WideLayout from "@src/components/layouts/WideLayout"
